+++ src/main/java/icbm/explosion/explosive/blast/BlastAntiGravitational.java
@@ -122,7 +122,7 @@ public class BlastAntiGravitational extends Blast
     }
 
     @Override
-    public float getEnergy()
+    public long getEnergy()
     {
         return 10000;
     }
